+++ src/ft_puthex_min.c
@@ -6,13 +6,13 @@
 /*   By: maborges <maborges@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/12 17:05:22 by maborges          #+#    #+#             */
-/*   Updated: 2024/12/16 18:43:43 by maborges         ###   ########.fr       */
+/*   Updated: 2025/01/24 15:38:14 by maborges         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_puthex_min(unsigned long hex)
+int	ft_puthex_min(unsigned int hex)
 {
 	char	*hexdigits;
 	int		count;
